+++ src/apps/js_generic/js_generic.cpp
@@ -332,7 +332,13 @@ namespace ccfapp
       js::register_request_body_class(ctx);
       js::populate_global_console(ctx);
       js::populate_global_ccf(
-        &txctx, transaction_id, receipt, nullptr, nullptr, ctx);
+        &txctx,
+        args.rpc_ctx.get(),
+        transaction_id,
+        receipt,
+        nullptr,
+        nullptr,
+        ctx);
 
       // Compile module
       std::string code = handler_script.value();
@@ -660,11 +666,6 @@ namespace ccfapp
             ds::openapi::path(document, key.uri_path),
             http_verb.value(),
             false);
-          LOG_INFO_FMT(
-            "Building OpenAPI for {} {}", key.verb.c_str(), key.uri_path);
-          const auto dumped = document.dump(2);
-          LOG_INFO_FMT(
-            "Starting from: {}", std::string(dumped.begin(), dumped.end()));
           if (!properties.openapi.empty())
           {
             for (const auto& [k, v] : properties.openapi.items())