+++ src/node/network_tables.h
@@ -22,7 +22,6 @@
 #include "secrets.h"
 #include "service.h"
 #include "service_map.h"
-#include "service_principals.h"
 #include "shares.h"
 #include "signatures.h"
 #include "snapshot_evidence.h"
@@ -82,8 +81,6 @@ namespace ccf
     UserCerts user_certs;
     UserInfo user_info;
 
-    ServicePrincipals service_principals;
-
     //
     // Node table
     //
@@ -133,7 +130,6 @@ namespace ccf
       jwt_public_signing_key_issuer(Tables::JWT_PUBLIC_SIGNING_KEY_ISSUER),
       user_certs(Tables::USER_CERTS),
       user_info(Tables::USER_INFO),
-      service_principals(Tables::SERVICE_PRINCIPALS),
       nodes(Tables::NODES),
       service(Tables::SERVICE),
       values(Tables::VALUES),