+++ tests/ca_certs.py
@@ -61,7 +61,7 @@ def test_cert_store(network, args):
         )
 
         stored_cert = json.loads(
-            primary.get_ledger_public_state_at(set_proposal.completed_seqno)[
+            network.get_ledger_public_state_at(set_proposal.completed_seqno)[
                 "public:ccf.gov.tls.ca_cert_bundles"
             ][raw_cert_name]
         )
@@ -74,7 +74,7 @@ def test_cert_store(network, args):
     remove_proposal = network.consortium.remove_ca_cert_bundle(primary, cert_name)
 
     assert (
-        primary.get_ledger_public_state_at(remove_proposal.completed_seqno)[
+        network.get_ledger_public_state_at(remove_proposal.completed_seqno)[
             "public:ccf.gov.tls.ca_cert_bundles"
         ][raw_cert_name]
         == None