+++ tests/governance_history.py
@@ -12,16 +12,13 @@ import http
 import base64
 import json
 from loguru import logger as LOG
+import suite.test_requirements as reqs
 
 
-def count_governance_operations(ledger):
+def check_operations(ledger, operations):
     LOG.debug("Audit the ledger file for governance operations")
 
     members = {}
-    verified_votes = 0
-    verified_proposals = 0
-    verified_withdrawals = 0
-
     for chunk in ledger:
         for tr in chunk:
             tables = tr.get_public_domain().get_tables()
@@ -34,7 +31,6 @@ def count_governance_operations(ledger):
                 governance_history_table = tables["public:ccf.gov.history"]
                 for member_id, signed_request in governance_history_table.items():
                     assert member_id in members
-
                     signed_request = json.loads(signed_request)
 
                     cert = members[member_id]
@@ -49,14 +45,19 @@ def count_governance_operations(ledger):
                     request_target_line = req.decode().splitlines()[0]
                     if "/gov/proposals" in request_target_line:
                         vote_suffix = "/ballots"
+                        elements = request_target_line.split("/")
                         if request_target_line.endswith(vote_suffix):
-                            verified_votes += 1
+                            op = (elements[-2], member_id.decode(), "vote")
                         elif request_target_line.endswith("/withdraw"):
-                            verified_withdrawals += 1
+                            op = (elements[-2], member_id.decode(), "withdraw")
                         else:
-                            verified_proposals += 1
+                            (proposal_id,) = tables["public:ccf.gov.proposals"].keys()
+                            op = (proposal_id.decode(), member_id.decode(), "propose")
 
-    return (verified_proposals, verified_votes, verified_withdrawals)
+                        if op in operations:
+                            operations.remove(op)
+
+    assert operations == set(), operations
 
 
 def check_all_tables_are_documented(ledger, doc_path):
@@ -69,20 +70,38 @@ def check_all_tables_are_documented(ledger, doc_path):
         for tr in chunk:
             table_names_in_ledger.update(tr.get_public_domain().get_tables().keys())
 
-    gov_tables_in_ledger = set(
-        [tn for tn in table_names_in_ledger if tn.startswith("public:ccf.gov.")]
+    public_table_names_in_ledger = set(
+        [tn for tn in table_names_in_ledger if tn.startswith("public:ccf.")]
     )
-    undocumented_tables = gov_tables_in_ledger - set(table_names)
-    LOG.info(undocumented_tables)
-    # Enable once Lua governance removal is complete
-    # assert undocumented_tables == set(), undocumented_tables
+    undocumented_tables = public_table_names_in_ledger - set(table_names)
+    assert undocumented_tables == set(), undocumented_tables
+
+
+@reqs.description("Check tables are documented")
+def test_tables_doc(network, args):
+    primary, _ = network.find_primary()
+    ledger_directories = primary.remote.ledger_paths()
+    ledger = ccf.ledger.Ledger(ledger_directories)
+    check_all_tables_are_documented(ledger, "../doc/audit/builtin_maps.rst")
+    return network
+
+
+@reqs.description("Test that all node's ledgers can be read")
+def test_ledger_is_readable(network, args):
+    primary, backups = network.find_nodes()
+    for node in (primary, *backups):
+        ledger_dirs = node.remote.ledger_paths()
+        LOG.info(f"Reading ledger from {ledger_dirs}")
+        ledger = ccf.ledger.Ledger(ledger_dirs)
+        for chunk in ledger:
+            for _ in chunk:
+                pass
+    return network
 
 
 def run(args):
-    # Keep track of how many propose, vote and withdraw are issued in this test
-    proposals_issued = 0
-    votes_issued = 0
-    withdrawals_issued = 0
+    # Keep track of governance operations that happened in the test
+    governance_operations = set()
 
     with infra.network.network(
         args.nodes, args.binary_dir, args.debug_nodes, args.perf_nodes, pdb=args.pdb
@@ -90,37 +109,40 @@ def run(args):
         network.start_and_join(args)
         primary, _ = network.find_primary()
 
-        ledger_directory = primary.remote.ledger_path()
-
-        ledger = ccf.ledger.Ledger(ledger_directory)
-        (
-            original_proposals,
-            original_votes,
-            original_withdrawals,
-        ) = count_governance_operations(ledger)
-
+        ledger_directories = primary.remote.ledger_paths()
         LOG.info("Add new member proposal (implicit vote)")
         (
             new_member_proposal,
             _,
             careful_vote,
         ) = network.consortium.generate_and_propose_new_member(
-            primary, curve=infra.network.ParticipantsCurve.secp256r1
+            primary, curve=infra.network.EllipticCurve.secp256r1
         )
-        proposals_issued += 1
+        member = network.consortium.get_member_by_local_id(
+            new_member_proposal.proposer_id
+        )
+        governance_operations.add(
+            (new_member_proposal.proposal_id, member.service_id, "propose")
+        )
 
         LOG.info("2/3 members accept the proposal")
         p = network.consortium.vote_using_majority(
             primary, new_member_proposal, careful_vote
         )
-        votes_issued += p.votes_for
+        for voter in p.voters:
+            governance_operations.add((p.proposal_id, voter, "vote"))
         assert new_member_proposal.state == infra.proposal.ProposalState.ACCEPTED
 
         LOG.info("Create new proposal but withdraw it before it is accepted")
         new_member_proposal, _, _ = network.consortium.generate_and_propose_new_member(
-            primary, curve=infra.network.ParticipantsCurve.secp256r1
+            primary, curve=infra.network.EllipticCurve.secp256r1
         )
-        proposals_issued += 1
+        member = network.consortium.get_member_by_local_id(
+            new_member_proposal.proposer_id
+        )
+        governance_operations.add(
+            (new_member_proposal.proposal_id, member.service_id, "propose")
+        )
 
         with primary.client() as c:
             response = network.consortium.get_member_by_local_id(
@@ -129,29 +151,20 @@ def run(args):
             infra.checker.Checker(c)(response)
         assert response.status_code == http.HTTPStatus.OK.value
         assert response.body.json()["state"] == ProposalState.WITHDRAWN.value
-        withdrawals_issued += 1
+        member = network.consortium.get_member_by_local_id(
+            new_member_proposal.proposer_id
+        )
+        governance_operations.add(
+            (new_member_proposal.proposal_id, member.service_id, "withdraw")
+        )
 
-    # Refresh ledger to beginning
-    ledger = ccf.ledger.Ledger(ledger_directory)
+        # Force ledger flush of all transactions so far
+        network.get_latest_ledger_public_state()
+        ledger = ccf.ledger.Ledger(ledger_directories)
+        check_operations(ledger, governance_operations)
 
-    (
-        final_proposals,
-        final_votes,
-        final_withdrawals,
-    ) = count_governance_operations(ledger)
-
-    assert (
-        final_proposals == original_proposals + proposals_issued
-    ), f"Unexpected number of propose operations recorded in the ledger (expected {original_proposals + proposals_issued}, found {final_proposals})"
-    assert (
-        final_votes == original_votes + votes_issued
-    ), f"Unexpected number of vote operations recorded in the ledger (expected {original_votes + votes_issued}, found {final_votes})"
-    assert (
-        final_withdrawals == original_withdrawals + withdrawals_issued
-    ), f"Unexpected number of withdraw operations recorded in the ledger (expected {original_withdrawals + withdrawals_issued}, found {final_withdrawals})"
-
-    ledger = ccf.ledger.Ledger(ledger_directory)
-    check_all_tables_are_documented(ledger, "../doc/audit/builtin_maps.rst")
+        test_ledger_is_readable(network, args)
+        test_tables_doc(network, args)
 
 
 if __name__ == "__main__":