+++ tests/infra/node.py
@@ -313,32 +313,16 @@ class Node:
                 f"Node {self.local_node_id} failed to join the network"
             ) from e
 
-    def get_ledger_public_state_at(self, seqno, timeout=3):
-        end_time = time.time() + timeout
-        while time.time() < end_time:
-            try:
-                ledger = ccf.ledger.Ledger(self.remote.ledger_path())
-                tx = ledger.get_transaction(seqno)
-                return tx.get_public_domain().get_tables()
-            except Exception:
-                time.sleep(0.1)
+    def get_ledger_public_tables_at(self, seqno):
+        ledger = ccf.ledger.Ledger(self.remote.ledger_paths())
+        assert ledger.last_committed_chunk_range[1] >= seqno
+        tx = ledger.get_transaction(seqno)
+        return tx.get_public_domain().get_tables()
 
-        raise TimeoutError(
-            f"Could not read transaction at seqno {seqno} from ledger {self.remote.ledger_path()}"
-        )
-
-    def get_latest_ledger_public_state(self, timeout=3):
-        end_time = time.time() + timeout
-        while time.time() < end_time:
-            try:
-                ledger = ccf.ledger.Ledger(self.remote.ledger_path())
-                return ledger.get_latest_public_state()
-            except Exception:
-                time.sleep(0.1)
-
-        raise TimeoutError(
-            f"Could not read latest state from ledger {self.remote.ledger_path()}"
-        )
+    def get_ledger_public_state_at(self, seqno):
+        ledger = ccf.ledger.Ledger(self.remote.ledger_paths())
+        assert ledger.last_committed_chunk_range[1] >= seqno
+        return ledger.get_latest_public_state()
 
     def get_ledger(self, include_read_only_dirs=False):
         """